refactor: separate core kanban files
