+++ apps/web/src/components/kanban/kanban-container.tsx
@@ -1,8 +1,9 @@
 import { DragDropContext, type OnDragEndResponder } from '@hello-pangea/dnd';
 import { type KanbanColumn } from '@utils/transform-to-column';
 import { useKanbanColumns } from 'src/hooks/useKanbanColumns';
-import { getInvolvedColumns, getMovingItem } from './core';
 import KanbanCol from './kanban-column';
+import { getInvolvedColumns } from './core/getInvolvedColumns';
+import { getMovingItemData } from './core/getMovingItemData';
 
 export default function JobsKanban({ jobColumns }: { jobColumns: KanbanColumn[] }) {
     const { columns, updateMovedItem } = useKanbanColumns(jobColumns);
@@ -10,11 +11,8 @@ export default function JobsKanban({ jobColumns }: { jobColumns: KanbanColumn[] 
     const onDragEnd: OnDragEndResponder = (result) => {
         const involvedColumns = getInvolvedColumns(columns, result);
         if (!involvedColumns) return;
-
-        const movingItem = getMovingItem(result, involvedColumns);
-        if (!movingItem) return;
-
-        updateMovedItem(movingItem.movingItem, movingItem.sourceJobs, movingItem.destinationJobs, involvedColumns).then(res => {
+        const movingItemData = getMovingItemData(result, involvedColumns);
+        updateMovedItem(movingItemData, involvedColumns).then(res => {
             // handle success state
         }).catch(err => {
             // handle error state
